Introducing Blueprint as schema defintion for models, for now.
Working migrations!

+++ lib/meta_orm/schema_comparer.rb
@@ -1,50 +1,64 @@
 # frozen_string_literal: true
 
-module MetaORM
+module MetaOrm
   class SchemaComparer
-    def self.generate_migration(model_class)
-      model_attributes = model_class.attributes_meta
-      table_columns = DB[model_class.table_name].columns
-      table_info = DB.schema(model_class.table_name)
+    def self.generate_migration(blueprint_class)
+      model_attributes = blueprint_class.attributes_meta
+      table_name = blueprint_class.table_name
 
-      changes = []
+      table_exists = DB.table_exists?(table_name)
       timestamp = Time.now.strftime("%Y%m%d%H%M%S")
-      migration_file = "db/migrations/#{timestamp}_create_#{model_class.table_name}.rb"
+      migration_file = "db/migrations/#{timestamp}_create_#{table_name}.rb"
+      changes = []
 
-      missing_columns = model_attributes.keys - table_columns
-      missing_columns.each do |column|
-        type = model_attributes[column][:type]
-        changes << "add_column :#{column}, :#{type}#{column_options(model_attributes[column])}"
-      end
+      if table_exists
+        puts "Updating table #{table_name}..."
 
-      extra_columns = table_columns - model_attributes.keys
-      extra_columns.each do |column|
-        changes << "remove_column :#{column}"
-      end
+        table_columns = DB[table_name].columns
+        table_info    = DB.schema(table_name)
 
-      table_info.each do |column, info|
-        column_name = column.to_sym
-        if model_attributes[column_name]
-          column_type = info[0]
-          if column_type.to_s != model_attributes[column_name][:type].to_s
-            changes << "change_column :#{column}, :#{model_attributes[column_name][:type]}"
+        missing_columns = model_attributes.keys - table_columns
+        missing_columns.each do |column|
+          type = model_attributes[column][:type]
+          changes << "add_column :#{column}, :#{type}#{column_options(model_attributes[column])}"
+        end
+
+        extra_columns = table_columns - model_attributes.keys
+        extra_columns.each do |column|
+          changes << "remove_column :#{column}"
+        end
+
+        table_info.each do |column, info|
+          column_name = column.to_sym
+          if model_attributes[column_name]
+            column_type = info[0]
+            if column_type.to_s != model_attributes[column_name][:type].to_s
+              changes << "change_column :#{column}, :#{model_attributes[column_name][:type]}"
+            end
           end
         end
+      else
+        puts "Table #{table_name} does not exist. Generating full create_table migration..."
+        changes << "create_table :#{table_name} do"
+        model_attributes.each do |column, meta|
+          changes << "  column :#{column}, :#{meta[:type]}#{column_options(meta)}"
+        end
+        changes << "end"
       end
 
       if changes.any?
         File.open(migration_file, 'w') do |f|
-          f.puts "class Migration#{timestamp}Create#{model_class.table_name.capitalize}"
-          f.puts "  def change"
-          changes.each do |change|
-            f.puts "    #{change}"
-          end
+          f.puts "Sequel.migration do"
+          f.puts "  change do"
+          changes.each { |line| f.puts "    #{line}" }
           f.puts "  end"
           f.puts "end"
         end
         puts "Migration file created: #{migration_file}"
+        migration_file
       else
-        puts "No changes detected for model: #{model_class.name}"
+        puts "No changes detected for model: #{blueprint_class.name}"
+        nil
       end
     end
